move CompletionStepView to presentation
