+++ presentation/src/main/java/org/sagebionetworks/research/presentation/inject/StepViewModule.java
@@ -35,9 +35,11 @@ package org.sagebionetworks.research.presentation.inject;
 import android.support.annotation.Nullable;
 
 import org.sagebionetworks.research.domain.step.StepType;
+import org.sagebionetworks.research.domain.step.implementations.CompletionStepBase;
 import org.sagebionetworks.research.domain.step.interfaces.Step;
 import org.sagebionetworks.research.presentation.mapper.DrawableMapper;
 import org.sagebionetworks.research.presentation.model.implementations.ActiveUIStepViewBase;
+import org.sagebionetworks.research.presentation.model.implementations.CompletionStepViewBase;
 import org.sagebionetworks.research.presentation.model.implementations.FormUIStepViewBase;
 import org.sagebionetworks.research.presentation.model.implementations.UIStepViewBase;
 import org.sagebionetworks.research.presentation.model.interfaces.StepView;
@@ -87,6 +89,20 @@ public abstract class StepViewModule {
     }
 
     @Provides
+    @IntoMap
+    @StepTypeKey(StepType.UI)
+    static InternalStepViewFactory provideUIStepFactory() {
+        return UIStepViewBase::fromUIStep;
+    }
+
+    @Provides
+    @IntoMap
+    @StepTypeKey(StepType.COMPLETION)
+    static InternalStepViewFactory provideCompletionStepFactory() {
+        return CompletionStepViewBase::fromCompletionStep;
+    }
+
+    @Provides
     static StepViewFactory provideStepViewFactory(final Map<String, InternalStepViewFactory> stepToFactoryMap,
             final DrawableMapper drawableMapper) {
         return (final Step step) ->
@@ -98,12 +114,5 @@ public abstract class StepViewModule {
                 return UIStepViewBase.fromUIStep(step, drawableMapper);
             }
         };
-    }
-
-    @Provides
-    @IntoMap
-    @StepTypeKey(StepType.UI)
-    static InternalStepViewFactory provideUIStepFactory() {
-        return UIStepViewBase::fromUIStep;
     }
 }
